removing domain manager for now
