+++ common/isolationgroup/defaultisolationgroupstate/global-api-handlers.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package defaultisolationgroup
+package defaultisolationgroupstate
 
 import (
 	"context"
